Fix parsing comments
At least for simple cases FluentMark can now find the closing --> "bracket" for an opening <!-- "bracket", even if it is in the same line.
If there is more after --> (or after --->) than a simple line ending, it is ignored, i.e. the whole line is classified as a comment (similar to the original behavior).