Use latest markdown-core version: 1.0.10-20251110-1607
No significant changes in markdown-core, just added an explicit dependency to plantuml in order to override the version from flexmark and make that version transitively visible by other projects depending on markdown-core.

Irrelevant change for FluentMark, since it uses plantuml through an Eclipse plug-in. We just use the same markdown-core version as in Bitbucket's renderer.